requirements.txt
openai>=0.26.4
rich>=13.3.1
pyyaml>=6.0
